@@ -1,8 +1,6 @@
-// +build !openbsd
+//go:build !openbsd
 
 package protector
 
 // Protect calls OS specific protections like pledge on OpenBSD
-func Protect() {
-	return
-}
+func Protect() {}
