@@ -1,4 +1,4 @@
-// +build openbsd
+//go:build openbsd
 
 package protector
 
@@ -6,5 +6,5 @@ import "golang.org/x/sys/unix"
 
 // Protect calls OS specific protections like pledge on OpenBSD
 func Protect() {
-	unix.PledgePromises("stdio rpath tty proc exec")
+	unix.PledgePromises("stdio dpath wpath rpath tty proc exec inet tmppath")
 }
